Fix DDG: allow using conversations and improve UX for incorrect usage
The vqd-4 changes for each continuation of the converation, thus
allowing stepping back or branching, but this contradicts the initial
authors assumption. The logic is as such: first status update gives us
first empty chat id, the successive chat queries send the new
conversation id, we just got to read it and return the new one to the
user.